fixning number format in native app
