+++ app/components/TreecounterGraphics/TreecounterGraphicsText.js
@@ -4,7 +4,7 @@ import TargetComment from './TargetComment';
 import ArrowButton from '../Common/ArrowButton';
 import { pot, tree, tree_outline } from '../../assets';
 import i18n from '../../locales/i18n.js';
-import { delimitNumbers } from '../../utils/utils';
+import { delimitNumbers, convertNumber } from '../../utils/utils';
 import PlantedDetails from './PlantedDetails';
 
 class TreecounterGraphicsText extends Component {
@@ -24,38 +24,6 @@ class TreecounterGraphicsText extends Component {
   //     .replace(/,/g, '');
   // }
 
-  convertNumber(n, d) {
-    if (isNaN(n) || undefined) {
-      return 0;
-    }
-    let x = ('' + n).length;
-    if (x > 12) {
-      let p = Math.pow;
-      d = p(10, d);
-      x -= x % 3;
-      return (
-        delimitNumbers(Math.round(n * d / p(10, x)) / d) +
-        [
-          '',
-          ' ' + i18n.t('label.Thousand'),
-          ' ' + i18n.t('label.Million'),
-          ' ' + i18n.t('label.Billion'),
-          ' ' + i18n.t('label.Trillion'),
-          ' ' + i18n.t('label.Quadrillion'),
-          ' ' + i18n.t('label.Quintillion')
-        ][x / 3]
-      );
-    } else if (x > 9) {
-      return delimitNumbers(n / 1000000000) + ' ' + i18n.t('label.Billion');
-    } else if (x > 6) {
-      return delimitNumbers(n / 1000000) + ' ' + i18n.t('label.Million');
-    } else if (x > 3) {
-      return delimitNumbers(n / 1000) + ' ' + i18n.t('label.Thousand');
-    } else {
-      return delimitNumbers(n);
-    }
-  }
-
   updateState(stateVal) {
     this.setState({ ifPlantedDetails: stateVal });
     this.props.onToggle(stateVal);
@@ -74,8 +42,8 @@ class TreecounterGraphicsText extends Component {
       }
     } = this.props;
     let dom;
-    const targetCount = this.convertNumber(target, 2);
-    const plantedCount = this.convertNumber(parseInt(planted), 2);
+    const targetCount = convertNumber(target, 2);
+    const plantedCount = convertNumber(parseInt(planted), 2);
     {
       dom = !this.state.ifPlantedDetails ? (
         <div className="svg-text-container">
@@ -125,7 +93,7 @@ class TreecounterGraphicsText extends Component {
               </span>
             </div>
             {this.props.trillion ||
-            this.convertNumber(parseInt(community), 2) == 0 ? null : (
+            convertNumber(parseInt(community), 2) == 0 ? null : (
               <div className="svg-text-container__row--col2">
                 <ArrowButton onToggle={e => this.updateState(e)} />
               </div>
@@ -134,8 +102,8 @@ class TreecounterGraphicsText extends Component {
         </div>
       ) : (
         <PlantedDetails
-          personal={this.convertNumber(parseInt(personal), 2)}
-          community={this.convertNumber(parseInt(community), 2)}
+          personal={convertNumber(parseInt(personal), 2)}
+          community={convertNumber(parseInt(community), 2)}
           type={type}
           onToggle={e => this.updateState(false)}
         />
